requirements.txt
pyramid
pyramid_debugtoolbar
gevent
pastegevent
